Update the WPF and syncfusion dependencies.
